Move `editGraph` to `editor/index.ts`.
